+++ packages/frontend/app/(dashboard)/course/[cid]/async_centre/components/modals/ConvertChatbotQToAnytimeQModal.tsx
@@ -1,5 +1,5 @@
 import { useMemo, useState, useEffect } from 'react'
-import { Modal, Input, Form, message, Checkbox, Tooltip } from 'antd'
+import { Modal, Input, Form, message, Checkbox, Tooltip, Spin } from 'antd'
 import { useUserInfo } from '@/app/contexts/userContext'
 import { useQuestionTypes } from '@/app/hooks/useQuestionTypes'
 import { QuestionTagSelector } from '../../../components/QuestionTagElement'
@@ -16,6 +16,8 @@ interface FormValues {
   questionText: string
   questionTypesInput: number[]
   refreshAIAnswer: boolean
+  setVisible: boolean
+  setAnonymous: boolean
 }
 
 interface ConvertChatbotQToAnytimeQModalProps {
@@ -34,6 +36,7 @@ const ConvertChatbotQToAnytimeQModal: React.FC<
   const [form] = Form.useForm()
   const [isLoading, setIsLoading] = useState(false)
   const courseFeatures = useCourseFeatures(courseId)
+  const authorCanSetVisible = courseFeatures?.asyncCentreAuthorPublic ?? false
   const [isGeneratingAbstract, setIsGeneratingAbstract] = useState(false)
 
   const userQuestionText = useMemo(() => {
@@ -162,6 +165,8 @@ const ConvertChatbotQToAnytimeQModal: React.FC<
           answerText: aiAnswer,
           questionAbstract: values.QuestionAbstract,
           status: asyncQuestionStatus.AIAnsweredNeedsAttention,
+          isAnonymous: values.setAnonymous,
+          authorSetVisible: authorCanSetVisible ? values.setVisible : false,
         },
         courseId,
       )
@@ -215,6 +220,10 @@ const ConvertChatbotQToAnytimeQModal: React.FC<
           layout="vertical"
           form={form}
           name="form_in_modal"
+          initialValues={{
+            setVisible: false,
+            setAnonymous: courseFeatures?.asyncCentreDefaultAnonymous ?? true,
+          }}
           clearOnDestroy
           onFinish={(values) => onFinish(values)}
         >
@@ -267,6 +276,30 @@ const ConvertChatbotQToAnytimeQModal: React.FC<
           <QuestionTagSelector questionTags={questionTypes} />
         </Form.Item>
       )}
+      {authorCanSetVisible && (
+        <Form.Item
+          name="setVisible"
+          label="Show Publicly?"
+          tooltip={
+            'Allows other students to see your question, and maybe even give you help via comments. Note that Course Staff can make questions public or private regardless of this setting.'
+          }
+          valuePropName="checked"
+          layout="horizontal"
+        >
+          <Checkbox />
+        </Form.Item>
+      )}
+      <Form.Item
+        name="setAnonymous"
+        label="Appear Anonymous?"
+        tooltip={
+          'If toggled, your name and avatar will not be shown with the question. Staff members will still see who you are.'
+        }
+        layout="horizontal"
+        valuePropName="checked"
+      >
+        <Checkbox />
+      </Form.Item>
       {courseFeatures?.asyncCentreAIAnswers && (
         <Tooltip
           placement="topLeft"
@@ -287,11 +320,6 @@ const ConvertChatbotQToAnytimeQModal: React.FC<
           </Form.Item>
         </Tooltip>
       )}
-      <div className="text-gray-500">
-        Only you and faculty will be able to see your question unless a faculty
-        member chooses to mark it public, in which case it will appear fully
-        anonymous to other students.
-      </div>
     </Modal>
   )
 }